Move globals into types directory
