+++ tsconfig.json
@@ -11,4 +11,8 @@
     "esModuleInterop": true,
     "resolveJsonModule": true,
   },
+  "typeRoots": [
+    "node_modules/@types",
+    "./types"
+  ]
 }
